+++ public/data/data.json
@@ -1,0 +1,14 @@
+[
+{"id" : 1, "username" : "yang" },
+{"id" : 2, "username" : "dong" },
+{"id" : 3, "username" : "seon" }
+]
+
+ 
+
+
+
+
+
+
+
