+++ src/pages/dongseon/Main/Aside.js
@@ -1,0 +1,15 @@
+const Aside = [
+  { id: 1, menulist: '소개' },
+  { id: 2, menulist: '도움말' },
+  { id: 3, menulist: '홍보센터' },
+  { id: 4, menulist: 'API' },
+  { id: 5, menulist: '채용정보' },
+  { id: 6, menulist: '개인정보처리방침' },
+  { id: 7, menulist: '약관' },
+  { id: 8, menulist: '위치' },
+  { id: 9, menulist: '인기계정' },
+  { id: 10, menulist: '해시태그' },
+  { id: 11, menulist: '언어' },
+];
+
+export default Aside;
